* origin/develop:
  ibv: increase default MaxMedium to 64KB
  UPCXX CI: Add issue312
  Update ChangeLog
  portable_inttypes: Be more trusting of modern language standards
  UPCXX CI: Apply "runnable" hack to _sympack prereq target
  UPCXX CI: Enable symPACK to spawn over udp-conduit
  UPCXX CI: Improve Fortran/symPACK handling
  UPCXX CI: Fix application of UPCXX_FC to CMake
  UPCXX CI: uniform application of UPCXX_EXTRA_FLAGS
  doc updates for pull request #331
  fixup: pmi-spawer: support for PMIx (PMI1 or PM...
  use fancier assert macros for PMI* API call results
  pmix malloc hack: make message more informative, check for all 4 entry points
  list x as a valid with-pmi-version value
  tidy up display of detected PMI version (if any) in configure output
  unify PMIx malloc hacks and fix missing variable declaration
  SQUASH: more recent PMIx uses calloc in header yoo
  pmi-spawer: support for PMIx (PMI1 or PMI2 modes)
  Update the configure logic to support --with-pmi-libs. I'm not sure I now have this right, though, as I can no longer run an application. I configure with: ``` ./configure --prefix=/home/common/gasnet/build/master --with-pmi-home=/home/common/pmix/build/master --with-pmi-version=x --enable-pmi --with-pmi-libs="-lpmix -Wl,-rpath=/home/common/pmix/build/master/lib" ``` and I get an error when trying to run: ``` $ prun -n 2 ./testgasnet ./testgasnet: symbol lookup error: ./testgasnet: undefined symbol: PMIx_Init ./testgasnet: symbol lookup error: ./testgasnet: undefined symbol: PMIx_Init $ ``` I've tried building static, but that generates an error when attempting to compile the tests.
  Update the PMIx support so it properly matches PMIx v2.x.
  UPCXX CI: Suppress a harmless warning from symPACK Cmake
  Suppress unused value warning in error abort code
  UPCXX CI: Restore phantom dependencies between build, install, test
  UPCXX CI: Add symPACK
  UPCXX CI: Fixups to cmake test
  ibv: add GASNetAMMaxMedium ident string
  ibv: implement --with-ibv-max-medium
  ibv: remove (debug only) timestamp from AM header
  aries: GASNetAMMaxMedium ident string
  UPCXX CI: add UPCXX_CONFIGURE_EXTRA support
  fixup: UPCXX CI: fix NDEBUG builds of cuda_vecadd
  UPCXX CI: add/correct support UPCXX_GASNET=srcdir
  fixup: ibv: refactor AM injection
  UPCXX CI: fix NDEBUG builds of cuda_vecadd
  UPCXX CI: fix arg quoting in launcher-mpi
  UPCXX CI: revise launcher-mpi
  UPCXX CI: rework environment handling
  UPCXX CI: repair extend-add recipe
  UPCXX CI: remove `GASNET{,_CONDUIT}` from env
  UPCXX CI: set UPCXX_NETWORK explicitly
  UPCXX CI: repair UPCXX_TARBALL_GASNET
  UPCXX CI: update Makefile comments
  UPCXX CI: remove now-unused UPCXX_BACKEND
  UPCXX CI: use installed launchers
  UPCXX CI: use installed makefile fragments
  UPCXX CI: remove remaining uses of nobs
  UPCXX CI: remove UPCXX_INSTALL_GASNET=1 support
  UPCXX CI: remove UPCXX_TESTMODE=nobs support
  fixup: ibv: refactor AM injection
  ibv: update README section "Graceful exits"
  ucx: update gasnet_ratomic_fwd.h
  ratomic: improve/update gasnet_ratomic_fwd.h
  ratomic: hoist handling of non signal-safe tools
  ucx: minor fix of prev commit
  ucx: fix handle all requests from recv queue while finalizing
  ucx: set `GASNETE_USING_REF_EXTENDED_GET/PUT` to zero
  ucx: eliminated duplications of `GASNETE_USING_REF_EXTENDED_GET/PUT`
  ucx: fixed recursive mutex acquisition for put_nb
  ucx: fix multi-threading support for atomics
  ucx: fix multi-threading support for extended support
  ucx: fixes for remote atomics
  ucx: repair support for --enable-threadinfo-opt
  ucx-conduit: update README
  ucx-conduit: update header file comments
  ucx-conduit: fix buid error
  ucx-conduit: added extended API: RDMA, Atomics
  ibv: document NPAM usage advice
  ibv: pass nargs to Commit
  NPAM: allow opt-in to passing numargs to Commit
  NPAM: move _tofree management "outward"
  NPAM: omit gasneti_init_srcdesc() when not needed
  ibv: drop _head_len from GASNETI_AM_SRCDESC_EXTRA
  ibv: suppress some maybe-uninitialized warnings
  ibv: lc_opt/gather-on-send handling for NPAM
  ibv: use gather-on-send via ODP when available
  ibv: implement GASNET_AM_GATHER_MIN
  ibv: implement NPAM Medium
  ibv: refactor AM injection
  ibv: remove AM-over-RDMA support
  testam: implement -max-step option
  testcore2: address NPAM coverage gaps (bug 4027)
  test.h: Fix an Open64 warning regression on gasnet tests
  THREADINFO: Fix a PGI C++ warning 550
  test(legacy)cxx: suppress warnings on PGI C++ 19.10/macos
  fixup: GASNET_POST_THREADINFO: fix a warning with pgc++/DEBUG
  fixup: gasnet_trace.h: Fix a pgc++ unused warning for !TRACE
  testcxx: fix some harmless pgc++ warnings
  testlegacycxx: fix a (harmless) typo discovered by pgc++ warnings
  test.h: fix some unused warnings from pgc++
  gasnet_trace.h: Fix a pgc++ unused warning for !TRACE
  GASNET_POST_THREADINFO: fix a warning with pgc++/DEBUG
  amudp: silence a harmless PGI C++ warning
  GASNETI_IDENT: minor improvement to reduce warnings
  configure: improve warning handling for pgc++
  bug 3810: workaround a bogus PGI warning on libupcr
  UPCXX CI: Add issue277 test for UPCXX PR141
  Require GNU Make 3.79 or newer
  testtools: fix a -Wsign-compare warning
  configure: report discovered perl version
  RAtomic: remove gasnet_ratomic_extra.h hook
  ibv: improve GASNET_USE_ODP documentation
  ibv: remove assertions of questionable value

+++ tests/testtools.c
@@ -1188,7 +1188,7 @@ void * thread_fn(void *arg) {
       uint64_t const share = MIN((unsigned)iters, (0xffffffffU / NUM_THREADS));
       uint64_t const one = 1;
       uint64_t const incrs[] = { one, one<<32, one + (one<<32) };
-      for (int i = 0; i < sizeof(incrs)/sizeof(incrs[0]); ++i) {
+      for (size_t i = 0; i < sizeof(incrs)/sizeof(incrs[0]); ++i) {
         uint64_t const incr = incrs[i];
         uint64_t const goal = NUM_THREADS * share * incr;
         uint64_t woncnt = 0;
